requirements.txt
asgiref==3.5.1
Django==4.0.4
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.48
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
